requirements.txt
openai>=1.30.0
librosa>=0.10.0
open-clip-torch>=2.24.0
opencv-python>=4.9.0
mediapipe>=0.10.0
moviepy>=2.0.0
click>=8.1.0
python-dotenv>=1.0.0
torch>=2.2.0
numpy>=1.26.0
soundfile>=0.12.0
